pyproject.toml
[project]
name = "agente-pessoal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.5.10",
    "anthropic>=0.86.0",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "sqlalchemy>=2.0.48",
]
